Moved class and left instructions for website entry
